Sync the browser url with the UrlBar url
The url is stored in the url fragment of the host browser. When the user
changes the hash location, the UrlBar location updates. When the user
changes the UrlBar location, the hash location updates.

It turned out that after changing the location hash in the tests, it was
necessary to yield to allow event handlers to run. The yield was
accomplished by waiting for 0 milliseconds.

An important lesson from this work is that it's important to manually
call `vm.$destroy()` on any component that needs cleanup when writing
tests.